pyproject.toml
[tool.poetry]
name = "pydantic-aws-cloudtrail"
version = "0.1.0"
description = ""
authors = ["Mike DuPont <jmikedupont2@gmail.com>"]
readme = "README.md"
packages = [
	 {include = "pydantic_aws_cloudtrail", from = "src"},
	 {include = "pydantic_metamodel", from = "src"},
	 {include = "json_schema", from = "src"},
]

[tool.poetry.scripts]
pydantic-aws-cloudtrail-generate-schema = "pydantic_aws_cloudtrail.__main__:main"
pct-fetch = "pydantic_aws_cloudtrail.fetch:main"
pmm-init = "pydantic_metamodel.model.meta.__init__:main"
pmm-generate-schema = "pydantic_metamodel.__main__:main"
jss-generate-schema = "json_schema.__main__:main"


[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.10.4"
pyyaml = "^6.0.2"
boto3 = "^1.35.90"
ruff = "^0.8.4"
datamodel-code-generator = "^0.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
